requirements.txt
iniconfig==2.0.0
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
pluggy==1.5.0
pytest==8.3.4
